backend/requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
falcon==3.1.1
h11==0.14.0
httptools==0.5.0
idna==3.4
kafka-python==2.0.2
openpyxl==3.1.1
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
